make updates work for more than amd64
previously arm64 and armhf would not get generated or mangled (for the
editions where those are enabled anyway - e.g. not release-lts) and as a
result their neon-desktop was woefully out of date.

update now is using bash instead of sh so we can use some nicer features
such as conditional assignment. it will now iterate all known possible
architectures and if a desktop-$arch file is present (that is: the arch
is in use for a given edition) it will wildcard the arch and mangle all
found files