disable modesetting hack
let's see how it goes. perhaps the original woes have gone away
or perhaps not. it'd be ideal if we can leave the hack out
altogether. if not, we should move to a config file based
approach as the package hack is proofing a stone to stumble over
for many users. (even though it works the way it does so it
makes it easy to switch away :|)